+++ raspi/src/wall_time.cpp
@@ -5,11 +5,11 @@ static auto _t_start = std::chrono::high_resolution_clock::now();
 uint32_t WallTime::millis()
 {
     auto t_now = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration<double, std::milli>(t_now - _t_start).count();
+    return static_cast<uint32_t>(std::chrono::duration<double, std::milli>(t_now - _t_start).count());
 }
 
 uint32_t WallTime::micros()
 {
     auto t_now = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration<double, std::micro>(t_now - _t_start).count();
+    return static_cast<uint32_t>(std::chrono::duration<double, std::micro>(t_now - _t_start).count());
 }
